[update] refactor
